+++ backend/app/controllers/api/reservations_controller.rb
@@ -7,8 +7,13 @@ module Api
     skip_before_action :verify_authenticity_token
 
     def create
-      @reservation = Reservation.new(requester: current_user, restaurant_id: reservation_params[:restaurant_id],
+      # @reservation = Reservation.new(requester: current_user, restaurant_id: reservation_params[:restaurant_id],
+      #                                date: reservation_params[:date], time_slot: reservation_params[:time_slot], party_size: reservation_params[:party_size])
+
+      @reservation = Reservation.new(restaurant_id: reservation_params[:restaurant_id],
                                      date: reservation_params[:date], time_slot: reservation_params[:time_slot], party_size: reservation_params[:party_size])
+      @reservation.user_id = current_user.id
+
       if @reservation.save!
         render :show
       else
@@ -17,20 +22,23 @@ module Api
     end
 
     def destroy
+      # puts reservation_params.inspect
       @reservation = Reservation.find_by(id: params[:id])
 
       if @reservation
         @reservation.destroy
       else
-        render json: ['Reservation not found'], status: :not_found
+        render json: [`Reservation not found`], status: :not_found
+
+        # render json: [`#{reservation_params}`]
       end
     end
 
-    private
+    # private
 
-    def reservation_params
-      params.require(:reservation).permit(:restaurant_id, :date, :time_slot, :party_size)
-    end
+    # def reservation_params
+    #   params.require(:reservation).permit(:restaurant_id, :date, :time_slot, :party_size)
+    # end
   end
 end
 # rubocop:enable Layout/LineLength
