+++ backend/config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
   namespace :api, defaults: { format: :json } do
     resources :users, only: %i[create show destroy]
     resource :session, only: %i[create destroy]
-    resource :reservations, only: %i[create destroy]
+    resources :reservations, only: %i[create destroy]
     resources :restaurants, only: %i[index create show] do
       resource :favorite, only: %i[create destroy]
       resource :reservations, only: %i[create destroy]
